refactor: move payment flow translation to app layer
