+++ test/unit/app/payments/translations.spec.ts
@@ -4,10 +4,11 @@ import { LedgerTransactionType } from "@domain/ledger"
 import {
   MissingPropsInTransactionForPaymentFlowError,
   NonLnPaymentTransactionForPaymentFlowError,
-  PaymentFlowFromLedgerTransaction,
 } from "@domain/payments"
 import { WalletCurrency } from "@domain/shared"
 import { PaymentInitiationMethod, SettlementMethod } from "@domain/wallets"
+
+import { PaymentFlowFromLedgerTransaction } from "@app/payments/translations"
 
 describe("PaymentFlowFromLedgerTransaction", () => {
   const satsAmount = toSats(20_000)
